diviso auth controller in profile controller
