+++ app/src/main/java/com/example/dietiestates25/controller/AuthController.kt
@@ -4,16 +4,13 @@ import android.content.Intent
 import android.util.Log
 import android.widget.TextView
 import androidx.appcompat.app.AppCompatActivity
-import androidx.fragment.app.Fragment
 import com.amplifyframework.auth.AuthUserAttribute
 import com.amplifyframework.auth.AuthUserAttributeKey
 import com.amplifyframework.auth.cognito.AWSCognitoAuthSession
-import com.amplifyframework.auth.cognito.result.AWSCognitoAuthSignOutResult
 import com.amplifyframework.auth.options.AuthSignUpOptions
 import com.amplifyframework.core.Amplify
 import com.example.dietiestates25.view.activity.HomeAgentActivity
 import com.example.dietiestates25.view.activity.HomeCustomerActivity
-import com.example.dietiestates25.view.activity.MainActivity
 
 class AuthController {
 
@@ -80,48 +77,6 @@ class AuthController {
         )
     }
 
-    fun signUpGestoreOrAgente(email: String, password: String, role: String, errorLabel: TextView, onSuccess: () -> Unit){
-        val attributes = listOf(
-            AuthUserAttribute(AuthUserAttributeKey.email(), email),
-            AuthUserAttribute(AuthUserAttributeKey.custom("custom:role"), role)
-        )
-
-        val options = AuthSignUpOptions.builder()
-            .userAttributes(attributes)
-            .build()
-
-        Amplify.Auth.signUp(
-            email,
-            password,
-            options,
-            {
-                errorLabel.visibility = TextView.INVISIBLE
-                Log.i("Amplify", "Sign up succeeded")
-                onSuccess()
-            },
-            {
-                errorLabel.visibility = TextView.VISIBLE
-                Log.e("Amplify", "Sign up failed", it)
-            }
-        )
-    }
-
-    fun updatePassword(oldPassword: String, newPassword: String, errorLabel: TextView, onSuccess: () -> Unit) {
-        Amplify.Auth.updatePassword(
-            oldPassword,
-            newPassword,
-            {
-                errorLabel.visibility = TextView.INVISIBLE
-                Log.i("Amplify", "Password updated successfully")
-                onSuccess()
-            },
-            { error ->
-                errorLabel.visibility = TextView.VISIBLE
-                Log.e("Amplify", "Password update failed", error)
-            }
-        )
-    }
-
     fun userIsSignedIn(activity: AppCompatActivity) {
         Amplify.Auth.fetchAuthSession(
             { session -> if (session.isSignedIn) {
@@ -130,52 +85,6 @@ class AuthController {
                 }
             }},
             { error -> Log.e("Auth", "Failed to fetch auth session", error) }
-        )
-    }
-
-    fun signOut(fragment: Fragment) {
-        Amplify.Auth.signOut { signOutResult ->
-            when(signOutResult) {
-                is AWSCognitoAuthSignOutResult.CompleteSignOut -> {
-                    // Sign Out completed fully and without errors.
-                    Log.i("AuthQuickStart", "Signed out successfully")
-                    comeBackToLogin(fragment)
-                }
-                is AWSCognitoAuthSignOutResult.PartialSignOut -> {
-                    // Sign Out completed with some errors. User is signed out of the device.
-                    signOutResult.hostedUIError?.let {
-                        Log.e("AuthQuickStart", "HostedUI Error", it.exception)
-                        // Optional: Re-launch it.url in a Custom tab to clear Cognito web session.
-
-                    }
-                    signOutResult.globalSignOutError?.let {
-                        Log.e("AuthQuickStart", "GlobalSignOut Error", it.exception)
-                        // Optional: Use escape hatch to retry revocation of it.accessToken.
-                    }
-                    signOutResult.revokeTokenError?.let {
-                        Log.e("AuthQuickStart", "RevokeToken Error", it.exception)
-                        // Optional: Use escape hatch to retry revocation of it.refreshToken.
-                    }
-                }
-                is AWSCognitoAuthSignOutResult.FailedSignOut -> {
-                    // Sign Out failed with an exception, leaving the user signed in.
-                    Log.e("AuthQuickStart", "Sign out Failed", signOutResult.exception)
-                }
-            }
-        }
-    }
-
-    fun fetchUserMail(callback: (String?) -> Unit) {
-        Amplify.Auth.fetchUserAttributes(
-            { attributes ->
-                val email = attributes.firstOrNull { it.key == AuthUserAttributeKey.email() }?.value
-                Log.i("AuthQuickstart", "User email: $email")
-                callback(email)
-            },
-            { error ->
-                Log.e("Auth", "Failed to fetch user attributes", error)
-                callback(null)
-            }
         )
     }
 
@@ -246,11 +155,5 @@ class AuthController {
     private fun goToHomeAgenti(activity: AppCompatActivity) {
         navigateTo(activity, HomeAgentActivity())
         activity.finish()
-    }
-
-    private fun comeBackToLogin(fragment: Fragment){
-        val intent = Intent(fragment.requireContext(), MainActivity::class.java)
-        fragment.startActivity(intent)
-        fragment.requireActivity().finish()
     }
 }
