+++ app/src/main/java/com/example/dietiestates25/view/fragment/RegistrationAgentFragment.kt
@@ -10,14 +10,14 @@ import android.widget.LinearLayout
 import android.widget.TextView
 import androidx.appcompat.app.AppCompatActivity
 import com.example.dietiestates25.R
-import com.example.dietiestates25.controller.AuthController
+import com.example.dietiestates25.controller.ProfileController
 
 class RegistrationAgentFragment : Fragment() {
-    private lateinit var authController: AuthController
+    private lateinit var profileController: ProfileController
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        authController = AuthController()
+        profileController = ProfileController()
     }
 
     override fun onCreateView(
@@ -49,7 +49,7 @@ class RegistrationAgentFragment : Fragment() {
     }
 
     private fun registerAgent(email: String, password: String, errorLabel: TextView) {
-        authController.signUpGestoreOrAgente(email, password, "AgenteImmobiliare", errorLabel) {
+        profileController.signUpGestoreOrAgente(email, password, "AgenteImmobiliare", errorLabel) {
             goBack()
         }
     }
