+++ app/src/main/java/com/example/dietiestates25/view/fragment/UpdatePasswordFragment.kt
@@ -10,14 +10,14 @@ import android.widget.LinearLayout
 import android.widget.TextView
 import androidx.appcompat.app.AppCompatActivity
 import com.example.dietiestates25.R
-import com.example.dietiestates25.controller.AuthController
+import com.example.dietiestates25.controller.ProfileController
 
 class UpdatePasswordFragment : Fragment() {
-    private lateinit var authController: AuthController
+    private lateinit var profileController: ProfileController
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        authController = AuthController()
+        profileController = ProfileController()
     }
 
     override fun onCreateView(
@@ -54,7 +54,7 @@ class UpdatePasswordFragment : Fragment() {
     }
 
     private fun updatePassword(oldPassword: String, newPassword: String, errorLabel: TextView) {
-        authController.updatePassword(oldPassword, newPassword, errorLabel) {
+        profileController.updatePassword(oldPassword, newPassword, errorLabel) {
             goBack()
         }
     }
